added chatroom context for props management
